Agrega vista Provincias. Modificaciones varias
Agrega los servicios para las tablas Localidad y Provincia. Ambas
derivan de una clase abstracta, para no tener que escribir tanto código
y para facilitar el manenimiento.

Agrega vista localidades. Al momento solo muestra el contenido de la
tabla de manera similar a la página Modelo.

Modificaciones en el contexto de la base de datos: se cambia el campo
Provincia(id) y Localidad(id_provincia) a entero sin signo.

+++ NeoRemiseria/Models/Localidad.cs
@@ -9,7 +9,7 @@ public partial class Localidad
 
     public string? Nombre { get; set; }
 
-    public byte? IdProvincia { get; set; }
+    public uint? IdProvincia { get; set; }
 
     public virtual Provincia? IdProvinciaNavigation { get; set; }
