+++ NeoRemiseria/Models/Provincia.cs
@@ -5,7 +5,7 @@ namespace NeoRemiseria.Models;
 
 public partial class Provincia
 {
-    public byte Id { get; set; }
+    public uint Id { get; set; }
 
     public string Nombre { get; set; } = null!;
